participants

+++ libcms/apps/participants_pages/administration/views.py
@@ -2,6 +2,7 @@
 from django.conf import settings
 from django.utils.translation import ugettext as _
 from django.shortcuts import render, get_object_or_404, redirect, HttpResponse
+from django.http import HttpResponseForbidden
 from guardian.decorators import permission_required_or_403
 from django.contrib.auth.decorators import login_required
 from common.pagination import get_page
@@ -9,17 +10,38 @@ from django.contrib.auth import login, REDIRECT_FIELD_NAME
 from django.utils.translation import to_locale, get_language
 
 from core.forms import LanguageForm
-from pages.models import Page, Content
+from participants_pages.models import Page, Content
 from forms import PageForm, ContentForm, get_content_form
+from participants.models import Library, LibraryContentEditor
+
+def get_cbs(library_node):
+    if library_node.parent_id:
+        return library_node.get_root()
+    else:
+        return library_node
+
+def check_owning(user, library):
+    if user.is_superuser:
+        return True
+    else:
+        if LibraryContentEditor.objects.filter(user=user, library=library).count():
+            return True
+        else:
+            return False
 
 #@permission_required_or_403('accounts.view_users')
-def index(request):
-    return redirect('pages:administration:pages_list')
-    #return render(request, 'pages/administration/index.html')
+def index(request, library_id):
+    return redirect('participants_pages:administration:pages_list', library_id=library_id)
+
 
 @login_required
-@permission_required_or_403('pages.add_page')
-def pages_list(request, parent=None):
+def pages_list(request, library_id, parent=None):
+
+    if not request.user.has_module_perms('participants_pages'):
+        return HttpResponseForbidden()
+    library = get_object_or_404(Library, id=library_id)
+
+
     if parent:
         parent = get_object_or_404(Page, id=parent)
 
@@ -36,42 +58,52 @@ def pages_list(request, parent=None):
     pages = [page['page'] for page in pages_dict.values()]
 
 
-    return render(request, 'pages/administration/pages_list.html', {
+    return render(request, 'participants_pages/administration/pages_list.html', {
         'parent': parent,
         'pages': pages,
         'pages_page': pages_page,
+        'library': library
     })
 
 @login_required
-@permission_required_or_403('pages.add_page')
-def create_page(request, parent=None):
+@permission_required_or_403('participants_pages.add_page')
+def create_page(request, library_id, parent=None):
+    library = get_object_or_404(Library, id=library_id)
+    cbs = get_cbs(library)
+    if not check_owning(request.user, cbs):
+        return HttpResponse(u'У Вас нет прав на создание страниц в этой ЦБС')
     if parent:
         parent = get_object_or_404(Page, id=parent)
 
     if request.method == 'POST':
-        page_form = PageForm(parent, request.POST, prefix='page_form')
-
+        page_form = PageForm(request.POST, prefix='page_form')
         if page_form.is_valid():
             page = page_form.save(commit=False)
             if parent:
                 page.parent = parent
 
-            if not request.user.has_perm('pages.public_page'):
+            if not request.user.has_perm('participants_pages.public_page'):
                 page.public = False
-
+            page.library = library
             page.save()
-            return redirect('pages:administration:create_page_content', page_id=page.id)
+            return redirect('participants_pages:administration:create_page_content', page_id=page.id, library_id=library_id)
     else:
         page_form = PageForm(prefix='page_form')
 
-    return render(request, 'pages/administration/create_page.html', {
+    return render(request, 'participants_pages/administration/create_page.html', {
         'parent': parent,
         'page_form': page_form,
+        'library': library
      })
 
 @login_required
-@permission_required_or_403('pages.change_page')
-def edit_page(request, id):
+@permission_required_or_403('participants_pages.change_page')
+def edit_page(request, library_id, id):
+    library = get_object_or_404(Library, id=library_id)
+    cbs = get_cbs(library)
+    if not check_owning(request.user, cbs):
+        return HttpResponse(u'У Вас нет прав на редактирование страницы в этой ЦБС')
+
     langs = []
     for lang in settings.LANGUAGES:
         langs.append({
@@ -86,18 +118,19 @@ def edit_page(request, id):
 
         if page_form.is_valid():
             page = page_form.save(commit=False)
-            if not request.user.has_perm('pages.public_page'):
+            if not request.user.has_perm('participants_pages.public_page'):
                 page.public = False
             page.save()
-            return redirect('pages:administration:pages_list')
+            return redirect('participants_pages:administration:pages_list', library_id=library_id)
 
     else:
         page_form = PageForm(prefix='page_form', instance=page)
 
-    return render(request, 'pages/administration/edit_page.html', {
+    return render(request, 'participants_pages/administration/edit_page.html', {
         'page': page,
         'langs': langs,
         'page_form': page_form,
+        'library': library
     })
 
 #@login_required
@@ -112,15 +145,23 @@ def edit_page(request, id):
 #    return redirect('pages:administration:pages_list')
 
 @login_required
-@permission_required_or_403('pages.delete_page')
-def delete_page(request, id):
+@permission_required_or_403('participants_pages.delete_page')
+def delete_page(request, library_id, id):
+    library = get_object_or_404(Library, id=library_id)
+    cbs = get_cbs(library)
+    if not check_owning(request.user, cbs):
+        return HttpResponse(u'У Вас нет прав на удаление страницы в этой ЦБС')
     page = get_object_or_404(Page, id=id)
     page.delete()
-    return redirect('pages:administration:pages_list')
+    return redirect('participants_pages:administration:pages_list', library_id=library_id)
 
 @login_required
-@permission_required_or_403('pages.add_page')
-def create_page_content(request, page_id):
+@permission_required_or_403('participants_pages.add_page')
+def create_page_content(request, library_id, page_id):
+    library = get_object_or_404(Library, id=library_id)
+    cbs = get_cbs(library)
+    if not check_owning(request.user, cbs):
+        return HttpResponse(u'У Вас нет прав на создание страницы в этой ЦБС')
     page = get_object_or_404(Page, id=page_id)
     if request.method == 'POST':
         content_form = ContentForm(request.POST, prefix='content_form')
@@ -132,19 +173,24 @@ def create_page_content(request, page_id):
 
             save = request.POST.get('save', u'save_edit')
             if save == u'save':
-                return redirect('pages:administration:edit_page', id=page_id)
+                return redirect('participants_pages:administration:edit_page', id=page_id, library_id=library_id)
             else:
-                return redirect('pages:administration:edit_page_content', page_id=page_id, lang=content.lang)
+                return redirect('participants_pages:administration:edit_page_content', page_id=page_id, lang=content.lang)
     else:
         content_form = ContentForm(prefix='content_form')
-    return render(request, 'pages/administration/create_page_content.html', {
+    return render(request, 'participants_pages/administration/create_page_content.html', {
         'page': page,
         'content_form': content_form,
+        'library': library
     })
 
 @login_required
-@permission_required_or_403('pages.change_page')
-def edit_page_content(request, page_id, lang):
+@permission_required_or_403('participants_pages.change_page')
+def edit_page_content(request, library_id,  page_id, lang):
+    library = get_object_or_404(Library, id=library_id)
+    cbs = get_cbs(library)
+    if not check_owning(request.user, cbs):
+        return HttpResponse(u'У Вас нет прав на редактирование страницы в этой ЦБС')
     lang_form = LanguageForm({'lang': lang})
     if not lang_form.is_valid():
         return HttpResponse(_(u'Language is not registered in system.') + _(u" Language code: ") + lang)
@@ -168,14 +214,16 @@ def edit_page_content(request, page_id, lang):
 
         save = request.POST.get('save', u'save_edit')
         if save == u'save':
-            return redirect('pages:administration:edit_page', id=page_id)
+            return redirect('participants_pages:administration:edit_page', id=page_id, library_id=library_id)
 
     else:
         content_form = ContentForm(prefix='content_form', instance=content)
-    return render(request, 'pages/administration/edit_page_content.html', {
+    return render(request, 'participants_pages/administration/edit_page_content.html', {
         'page': page,
         'content': content,
         'content_form': content_form,
+        'library': library
+
     })
